Rename 'descriptor/file.c' to 'descriptor/regular_file.c'
