+++ src/main/host/syscall/fileat.c
@@ -15,7 +15,7 @@
 
 #include "lib/logger/logger.h"
 #include "main/host/descriptor/descriptor.h"
-#include "main/host/descriptor/file.h"
+#include "main/host/descriptor/regular_file.h"
 #include "main/host/process.h"
 #include "main/host/syscall/kernel_types.h"
 #include "main/host/syscall/protected.h"
