+++ src/main/host/syscall/poll.c
@@ -13,7 +13,7 @@
 #include "main/host/descriptor/descriptor.h"
 #include "main/host/descriptor/descriptor_types.h"
 #include "main/host/descriptor/epoll.h"
-#include "main/host/descriptor/file.h"
+#include "main/host/descriptor/regular_file.h"
 #include "main/host/process.h"
 #include "main/host/status.h"
 #include "main/host/syscall/protected.h"
